vdp can draw some sprites

+++ fns.h
@@ -16,4 +16,4 @@ uint8_t vdpstatus(void);
 uint8_t vdpdataport(void);
 uint8_t vdphcounter(void);
 uint8_t vdpvcounter(void);
-void cramwrite(uint16_t, uint16_t);
+void cramwrite(uint16_t, uint8_t);
